refactor: move to foxhole.eval
this just makes everything more compact, you know?
