--- eval/queries.txt
@@ -1,1 +1,0 @@
-python library type checker
